+++ src/ayame/ayame_client.cpp
@@ -177,7 +177,7 @@ void AyameClient::CreatePeerConnection() {
 
   rtc_config.servers = ice_servers_;
   connection_ = manager_->CreateConnection(rtc_config, this);
-  manager_->InitTracks(connection_.get());
+  manager_->InitTracks(connection_.get(), config_.direction);
 }
 
 void AyameClient::Close() {